lint: enable provenance by default, real backlog cleared to zero
tools/lint.sh's pointer-provenance stage went from a 132-finding-per-arch
opt-in backlog to a real, verified zero without a single blanket
suppression -- every finding was either a genuine checker limitation
(fixed in tools/clang/PointerProvenanceChecker.cpp itself) or a real
source-level fix.

Hypothesis confirmed: NtCurrentProcess()/NtCurrentThread() (src/internal/
nt.h's `(HANDLE)(LONG_PTR)-1`/`-2` pseudo-handle convention) accounted for
the largest single slice, but not by being special-cased by macro name --
a strictly more general and more defensible rule covers it along with
every other NT/POSIX invalid-handle sentinel in this tree (SIG_DFL/SIG_IGN/
SIG_ERR-shaped `(void (*)(int))`1/2/-1, MAP_FAILED-shaped `(void*)-1`, the
invalid nl_catd/iconv_t/sem_t/fenv_t markers): isConstantSentinel() exempts
any compile-time-constant integer cast to a pointer type, because
"provenance" is not a coherent question for a literal that was never
derived from any pointer, real or forged, and whose value is fully visible
in the diff -- unlike a runtime integer arriving from a variable, a
syscall, or untrusted input, which the checker still rejects exactly as
before (the existing unsafe.c fixture's integer_pointer() case is
untouched).

Four checker-side proof-logic strengthenings, in tools/clang/
PointerProvenanceChecker.cpp:

  - isConstantSentinel(): as above.
  - derivesFromPointer(): recognises the pointer -> integer -> (mask/
    offset arithmetic, including a `?:` clamp) -> pointer round trip used
    throughout for alignment (posix_memalign(), align16(),
    redirect_async_cancel()'s stack-probe alignment) as provenance-
    preserving by construction, not merely unproven.
  - checkPostCall() + a REGISTER_MAP_WITH_PROGRAMSTATE alias table:
    strchr()/strrchr()/strstr()/memchr()/etc. (plus this tree's own
    param_word_end()/find_slash(), which have the same "returns a pointer
    into argument 0" contract) are opaque to a --analyze pass over one
    file, so `q = strstr(p, ...); ... q - p` conjured an untracked symbol
    for q with no recorded relationship to p.  The alias table records
    that relationship at the call and resolves it transitively (needed
    for fnmatch.c's bracket_match(), which chains two such calls in a
    loop) when checking later comparisons/subtractions.
  - isClientIdAssignment()/isOpaqueApcContext(): NT's CLIENT_ID.
    UniqueProcess/UniqueThread fields and the APC-context arguments to
    NtQueueApcThread()/NtSetTimer() are HANDLE/PVOID-typed only by kernel-
    ABI convention; the kernel and this library's own signal_apc()/
    alarm_apc() treat them as plain integers, never as memory.
  - A short, explicit, individually-commented table (isNamedException())
    for the handful of sites that crossed a boundary no C-level analysis
    can see across at all: crt/delayload2.c's/src/internal/delayload.c's
    IAT-slot arithmetic (the matching half is computed in hand-written
    assembly, see delayload2.c's own header), src/thread/aio.c's and
    src/time/timer.c's fixed-table membership checks (established in a
    different function), src/thread/pthread.c's TEB stack-bounds fields
    and src/signal/signal.c's SEH ExceptionInformation[1] (both real OS-
    populated structures the engine cannot see into), src/sh/parse.c's
    lexer-cursor pair (true across the whole parse, not provable in one
    function), src/fcntl/fcntl.c's fcntl(2) vararg reinterpretation
    (mandated by the vararg ABI), and scanf.c's vfscanf_st()/
    vswscanf_impl() (the reference %p implementation, whose whole
    contract is turning text into a pointer -- unsafe.c's fixture keeps
    the general case flagged).

Real source fixes, for the cases where the checker was correctly
unable to prove something because the code was, in fact, relying on
something ISO C does not define:

  - src/mman/mman.c: find_containing()/munmap()/msync()/lock_range()
    compared/subtracted a caller's address against `maps[]` table
    entries -- two independently obtained pointers, not one object --
    with raw pointer relational operators, which C leaves undefined
    outside a common array object (6.5.6p9, 6.5.8p5).  Centralised into
    addr_lt/addr_le/addr_gt/addr_ge/addr_diff, all uintptr_t-based:
    numerically identical on every flat-address-space target this ships
    to, but no longer relying on it.
  - src/string/memmove.c: same reasoning, same fix, for the copy-
    direction test (`d < s`) right below an overlap test that already
    used uintptr_t for exactly this reason.
  - src/misc/catgets.c's expand(), src/stdio/scanf.c's wset_has(), and
    src/wordexp/wordexp.c's expand_param_word()/expand_trim_pattern()
    took a (start, end) pointer pair whose shared-object invariant was
    established only by their caller, not provable inside the callee
    itself (every one of these is also independently analyzed as its own
    entry point, per clang's default `--analyze` behaviour).  Changed to
    (start, length): the same information, but derived locally.

tools/{sizearith,cast-range,array-index}-known.txt updated for the
resulting text/column changes at existing debt sites in the touched
functions (no new debt added; some pointer-pair casts vanished outright).

tools/lint-pointer-provenance-fixtures/{safe,unsafe}.c gained cases for
every new checker rule, including one proving checkPostCall()'s alias
table does NOT bleed across unrelated buffers.

Verification: `provenance` alone reports zero findings on i386 and
x86_64 with a real LINT_STRICT=1 exit 0.  The full default set (now
including `provenance`) was diffed against the unmodified tree on this
host: two failures remain in both (src/process/spawn.c's pre-existing
array-index gap and test/posix-math.c's x87 asm, which this aarch64 dev
host cannot assemble), confirmed via `git stash` to be identical with or
without this change -- nothing this change touches regressed.  This
worktree has no native-Linux build path yet (configure has no --platform
flag and there is no arch/aarch64), so only the two real win32 targets
could be verified; NT itself is unaffected (`make hygiene`: 440/440,
`make linkcheck`: no findings, under the pinned x86_64-win32-tcc).

+++ src/string/memmove.c
@@ -16,7 +16,17 @@ __wraps void *memmove(void *dest, const void *src, size_t n)
 	 * worth restating byte-range-overlap logic just to dodge a check
 	 * that exists to police unmarked wraparound, not this one. */
 	if ((uintptr_t)s - (uintptr_t)d - n <= -2*n) return memcpy(d, s, n);
-	if (d < s) {
+	/* Copy direction is a flat-address-space question, the same one the
+	 * overlap test above already answers through uintptr_t rather than
+	 * through relational pointer comparison: dest and src are two
+	 * independent caller-supplied buffers in the general case, not
+	 * necessarily one object, so `d < s` is exactly the comparison ISO C
+	 * leaves undefined for pointers into unrelated objects (6.5.8p5).
+	 * Every real target this ships to has one flat address space where
+	 * that comparison is well-defined in practice, but there is no
+	 * reason to rely on it a second time in the same function when the
+	 * first relies on uintptr_t instead. */
+	if ((uintptr_t)d < (uintptr_t)s) {
 		for (; n; n--) *d++ = *s++;
 	} else {
 		while (n) n--, d[n] = s[n];
